requirements.txt
fastapi
uvicorn
python-multipart
pillow
opencv-python
openai
python-dotenv
requests
numpy